feat: remove new credentialtyperegistry and use credentialrepository instead
On-behalf-of: SAP <matthias.bruns@sap.com>
