+++ bindings/go/plugin/manager/registries/credentialrepository/registry.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"log/slog"
 	"os"
 	"os/exec"
 	"sync"
@@ -21,20 +22,24 @@ func NewCredentialRepositoryRegistry(ctx context.Context) *RepositoryRegistry {
 		ctx:                                 ctx,
 		capabilities:                        make(map[string]credentialsv1.CapabilitySpec),
 		registry:                            make(map[runtime.Type]mtypes.Plugin),
+		credentialRegistry:                  make(map[runtime.Type]mtypes.Plugin),
 		constructedPlugins:                  make(map[string]*constructedPlugin), // running plugins
 		consumerTypeRegistrations:           make(map[runtime.Type]runtime.Type),
 		internalCredentialRepositoryPlugins: make(map[runtime.Type]credentials.RepositoryPlugin),
 		scheme:                              runtime.NewScheme(),
+		credentialTypeScheme:                runtime.NewScheme(),
 	}
 }
 
 // RepositoryRegistry holds all plugins that implement capabilities corresponding to RepositoryPlugin operations.
 type RepositoryRegistry struct {
-	ctx          context.Context
-	mu           sync.Mutex
-	capabilities map[string]credentialsv1.CapabilitySpec
-	registry     map[runtime.Type]mtypes.Plugin
-	scheme       *runtime.Scheme
+	ctx                  context.Context
+	mu                   sync.Mutex
+	capabilities         map[string]credentialsv1.CapabilitySpec
+	registry             map[runtime.Type]mtypes.Plugin
+	scheme               *runtime.Scheme
+	credentialRegistry   map[runtime.Type]mtypes.Plugin
+	credentialTypeScheme *runtime.Scheme
 
 	constructedPlugins        map[string]*constructedPlugin // running plugins
 	consumerTypeRegistrations map[runtime.Type]runtime.Type
@@ -44,6 +49,19 @@ type RepositoryRegistry struct {
 
 func (r *RepositoryRegistry) RepositoryScheme() *runtime.Scheme {
 	return r.scheme
+}
+
+func (r *RepositoryRegistry) GetCredentialTypeScheme() *runtime.Scheme {
+	return r.credentialTypeScheme
+}
+
+// Register merges a pre-built scheme of built-in credential types into the registry.
+// Call this during startup to make built-in types (e.g. OCICredentials, HelmHTTPCredentials)
+// known before any plugin is loaded.
+func (r *RepositoryRegistry) Register(scheme *runtime.Scheme) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	r.credentialTypeScheme.MustRegisterScheme(scheme)
 }
 
 // AddPlugin takes a plugin discovered by the manager and adds it to the stored plugin registry.
@@ -70,7 +88,41 @@ func (r *RepositoryRegistry) AddPlugin(plugin mtypes.Plugin, spec runtime.Typed)
 		r.registry[typ.Type] = plugin
 	}
 
+	for _, typ := range capability.CustomCredentialTypes {
+		if v, ok := r.credentialRegistry[typ.Type]; ok {
+			return fmt.Errorf("plugin for type %v already registered with ID: %s", typ.Type, v.ID)
+		}
+		r.credentialRegistry[typ.Type] = plugin
+	}
+	if err := r.registerCustomCredentialTypes(capability); err != nil {
+		return fmt.Errorf("failed to register custom credential types: %w", err)
+	}
+
 	return nil
+}
+
+func (r *RepositoryRegistry) registerCustomCredentialTypes(capability credentialsv1.CapabilitySpec) error {
+	var errs []error
+	for _, t := range capability.CustomCredentialTypes {
+		typed := &runtime.Raw{}
+		typed.SetType(t.Type)
+		allTypes := append([]runtime.Type{t.Type}, t.Aliases...)
+		conflict := false
+		for _, alias := range allTypes {
+			if r.credentialTypeScheme.IsRegistered(alias) {
+				errs = append(errs, fmt.Errorf("credential type %s already registered", alias))
+				conflict = true
+			}
+		}
+		if conflict {
+			continue
+		}
+		if err := r.credentialTypeScheme.RegisterWithAlias(typed, allTypes...); err != nil {
+			slog.ErrorContext(r.ctx, "failed to build scheme for plugin credential type", "type", t.Type, "error", err)
+			errs = append(errs, err)
+		}
+	}
+	return errors.Join(errs...)
 }
 
 func (r *RepositoryRegistry) GetPlugin(ctx context.Context, spec runtime.Typed) (credentials.RepositoryPlugin, error) {
